requirements.txt
Flask==2.3.3
SQLAlchemy>=2.0.31
Flask-SQLAlchemy==3.0.5
psycopg2
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT
blueprint